@@ -1,9 +1,13 @@
-import { Injectable, signal } from '@angular/core';
+import { Injectable, computed, effect, model, output, signal } from '@angular/core';
 import { BehaviorSubject } from 'rxjs';
 
 @Injectable({ providedIn: 'root' })
 export class CounterService {
   counter = signal(0);
+  futureCounter = computed(() => this.counter() + 1);
+  modelExample = model('');
+  outputExample = output<string>();
+
   serviceProp: number = 0;
   servicePropStream: BehaviorSubject<number> = new BehaviorSubject(0);
   increment() {
@@ -12,5 +16,13 @@ export class CounterService {
   incrementServiceProp() {
     this.serviceProp += 1;
     this.servicePropStream.next(this.serviceProp);
+  }
+
+  constructor() {
+    effect(() => {
+      console.log('counter: ', this.counter());
+    });
+
+    this.outputExample.emit('New Value');
   }
 }
